backend/requirements.txt
django>=6.0,<6.1
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-filter>=25.2,<26
psycopg[binary]>=3.2,<3.3
django-cors-headers>=4.7,<4.8
